+++ kotonebot/core/__init__.py
@@ -4,6 +4,7 @@ from .entities.ocr import OcrPrefab, OcrFindKargs
 from .entities.base import GameObject
 from .entities.base import GameObjectType
 from .entities.compound import AnyOf
+from .bot import *  # noqa: F403
 
 __all__ = [
     'Prefab', 'TemplateMatchPrefab', 'OcrPrefab',