+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Library:
 1. [feat] 新增多点触控接口，目前仅 nemu_ipc 实现了多点触控。
 2. [feat] 优化库 import 速度，从 2.9 秒降至 0.92 秒，提升了 68%。
 3. [feat] **BREAKING** 移除了内置的配置实体类与管理，以及 Context 上的 config 对象。
+4. [feat] 新增了新的 KotoneBot 任务调度类，位于 kotonebot.core.bot 内。原有旧对象 kotonebot.backend.bot 处于废弃状态，将与若干个版本后移除。
 
 ## v0.7.0
 Library: